Migrate network
