+++ fw/network/connection.go
@@ -1,0 +1,7 @@
+package network
+
+type Connection struct {
+	ClientIP      string
+	RequestedHost string
+	Protocol      string
+}
