+++ fw/network/network.go
@@ -1,0 +1,7 @@
+package network
+
+import "net/http"
+
+type Network interface {
+	FromHTTP(request *http.Request) Connection
+}
